+++ packages/react-core/src/operations/groupby.d.ts
@@ -1,9 +1,0 @@
-import { AggregationTypes } from './aggregation/AggregationTypes';
-import { GroupByFeature } from '../types';
-
-export function groupValuesByColumn(
-  data: [],
-  valuesColumn: string,
-  keysColumn: string,
-  operation: AggregationTypes
-): GroupByFeature;