+++ packages/react-core/src/types.d.ts
@@ -1,30 +1,35 @@
-import { TILE_FORMATS } from '@deck.gl/carto'
-import { AggregationTypes } from './operations/aggregation/AggregationTypes';
+import { TILE_FORMATS } from '@deck.gl/carto';
+import { AggregationTypes } from './operations/constants/AggregationTypes';
 import { Geometry } from 'geojson';
 
-export type AggregationFunctions = {
-  [AggregationTypes.COUNT]: Function,
-  [AggregationTypes.MIN]: Function,
-  [AggregationTypes.MAX]: Function,
-  [AggregationTypes.SUM]: Function,
-  [AggregationTypes.AVG]: Function
-}
+export type AggregationFunctions = Record<
+  AggregationTypes,
+  (
+    values: Record<string, unknown>[],
+    keys?: string | string[],
+    joinOperation?: AggregationTypes
+  ) => {}
+>;
 
-export type GroupByFeature = {
-  name: string,
-  value: number
-}[] | [];
+export type GroupByFeature =
+  | {
+      name: string;
+      value: number;
+    }[]
+  | [];
 
 export type HistogramFeature = number[] | [];
+
+export type ScatterPlotFeature = [number, number][];
 
 export type Viewport = [number, number, number, number];
 
 export type TileFeatures = {
-  tiles: any, // TODO: add proper deck.gl type
-  viewport: Viewport,
-  geometry?: Geometry,
-  uniqueIdProperty?: string,
-  tileFormat: TILE_FORMATS
-}
+  tiles: any; // TODO: add proper deck.gl type
+  viewport: Viewport;
+  geometry?: Geometry;
+  uniqueIdProperty?: string;
+  tileFormat: TILE_FORMATS;
+};
 
-export type TileFeaturesResponse = Record<string, unknown>[] | []
+export type TileFeaturesResponse = Record<string, unknown>[] | [];